requirements.txt
flask==2.3.3
pillow==10.0.0
opencv-python==4.8.0.76
numpy==1.25.2
requests==2.31.0
werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0